Refactor upload file actions
