+++ .releaserc
@@ -1,4 +1,12 @@
 {
-  "verifyConditions": ["@semantic-release/changelog", "@semantic-release/npm", "@semantic-release/git"],
-  "prepare": ["@semantic-release/changelog", "@semantic-release/npm", "@semantic-release/git"]
+  "verifyConditions": [
+    "@semantic-release/changelog",
+    "@semantic-release/npm",
+    "@semantic-release/git"
+  ],
+  "prepare": [
+    "@semantic-release/changelog",
+    "@semantic-release/npm",
+    "@semantic-release/git"
+  ]
 }
