+++ src/boards/index.js
@@ -1,8 +1,5 @@
 module.exports = {
   createBoard,
   findBoard,
-  // addFilesToBoard,
   addLinksToBoard,
-  // getFileUploadURLToBoard,
-  // completeFileUploadToBoard,
 } = require('./actions');
